Extract abstract IniDocumentBase from IniDocument
Introduce IniDocumentBase as the shared, read-oriented INI document model
(global section, ordered named sections, name lookup) with mutation reserved
to protected core methods. IniDocument becomes a sealed subclass that
re-exposes AddSection/GetOrAddSection/RemoveSection publicly, preserving its
exact public surface. This prepares a clean read-only base for the
configuration layer to inherit without re-leaking INI mutation.

All 668 Bodu.Text.Formats tests pass unchanged.
